auth, message posts

+++ base/views.py
@@ -1,7 +1,13 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from .models import Room, Topic, User, Message
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
+
 
 #authentication => session based by default
 
@@ -12,6 +18,55 @@ from .forms import RoomForm
 #   { 'id':2, 'name': 'Design with me'  },
 #   { 'id':3, 'name': 'lets learn nodejs bruv'  }
 # ]
+
+def loginPage(request):
+
+  page = 'login'
+
+  if request.user.is_authenticated:
+    messages.info(request, 'You are already logged in')
+    return redirect('Home')
+
+  if(request.method == 'POST'):
+    username = request.POST.get('username').lower()
+    password = request.POST.get('password')
+
+    try:
+      user = User.objects.get(username=username)
+    except:
+      messages.error(request, 'user does not exist!')
+    user = authenticate(request, username=username, password=password)
+    if user is not None:
+      login(request, user)  
+      return redirect('Home')
+    else:
+      messages.error(request, 'Invalid credentials!') #username or password does not exist
+      
+
+  context = {'page': page}
+  return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+  logout(request)
+  return redirect('Home')
+
+def registerUser(request):
+  form = UserCreationForm()
+  # context = {'form': form}
+
+  if request.method == 'POST':
+    form = UserCreationForm(request.POST)
+    if form.is_valid:
+      user = form.save(commit=False)
+      user.username = user.username.lower()
+      user.save()
+      login(request, user)
+      return redirect('Home')
+    else:
+      messages.error(request, 'an error occured during registration')
+  else:
+    messages.error(request, 'request not "POST"')    
+  return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
   q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -31,13 +86,19 @@ def home(request):
 
 def room(request, pk):
   room = Room.objects.get(id=pk)
-  # for i in rooms:
-  #   if i['id'] == int(pk):
-  #     room = i
-  context = {'room': room}    
+  room_messages = room.message_set.all().order_by('-created')
+  if request.method == 'POST':
+    message = Message.objects.create(
+      user = request.user,
+      room = room,
+      body = request.POST.get('body')
+    )
+    return redirect('room', pk=room.id)
+  context = {'room': room, 'room_messages': room_messages}    
 
   return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     if request.method == 'POST':
       form = RoomForm(request.POST)
@@ -53,9 +114,13 @@ def createRoom(request):
     context = { 'form': form }
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    return HttpResponse("you a'int allowed here")
 
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
@@ -69,8 +134,12 @@ def updateRoom(request, pk):
   context = {'room': form}
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
   room = Room.objects.get(id=pk)
+
+  if request.user != room.host and not request.user.is_superuser:
+    return HttpResponse("you a'int allowed here")
 
   if request.method == 'POST':
     room.delete()
@@ -78,4 +147,4 @@ def deleteRoom(request, pk):
   return render(request, 'base/delete.html', {'obj': room})
 
 
-#22517
+#31348
